Completely Remade Arc System Shader
So with Granblue on the horizon, I wanted to remake my Arc System shader completely from the ground up. Much better colors, actually added in the vertex colors because I was lazy before, and finalized my system that I started working on to automatically switch between a fake light direction if there is no world light source, and the real light direction if there is one.
I've scrapped the other versions as they are basically worthless now.
Also had updated some other misc shaders, but that really wasn't that important.