Big overhaul of UI, added back in outlines with a working toggle, removed and standardized keywords
Really big overhaul of the shader GUI from the help of Mochie and Morioh
Added back in a built in outline that just toggles the visibility on and off. You will still take the extra drawcall, but hey you can turn it off and on now you lazy lads
Removed every single custom keyword and swapped them out with unity standard keywords or with enums instead. Will remove that dreded "remove custom keywords" button if you are uploading to VRChat

Also did some work on custom functions, but not really that important